backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.0
# Auth & Security
python-jose[cryptography]
passlib[bcrypt]
python-multipart